[Refactor][Kernel] Rename kernels/conv/ to kernels/convolution/
Match the ops/convolution.py naming convention. Update all internal
and external imports.

Part 2 of #924.
